+++ tests/commander/test_commander_async.py
@@ -2,6 +2,9 @@ import pytest
 import threading
 import time
 
+from unittest.mock import Mock, AsyncMock
+
+
 from agere.commander._commander import (
     PASS_WORD,
     CommanderAsync,
@@ -192,4 +195,102 @@ def test_wait_for_exit(job_add, commander: CommanderAsync):
     run_thread_2.join()
     assert not commander.running_status
     assert exit_code is None
+    assert manipulate[0] == 2
+
+
+def test_commander_is_empty(commander, job_add):
+    # Setup
+    manipulate = [0]
+    job = job_add(manipulate)
+
+    # Assert
+    assert commander.is_empty()
+
+    # Action
+    threading.Thread(target=commander.run_auto, args=(job,)).start()
+    commander.wait_for_exit()
+
+    # Assert
+    assert manipulate[0] == 1
+    assert commander.is_empty()
+
+
+async def test_callback_handle(commander: CommanderAsync):
+    # Setup
+    mock_callback = Mock()
+    mock_task_node = Mock()
+    mock_callback._task_node = mock_task_node
+    mock_fn_1 = Mock()
+    mock_fn_2 = Mock()
+    mock_fn_3 = Mock()
+    mock_fn_4 = Mock()
+    mock_fn_5 = AsyncMock()
+    mock_fn_6 = AsyncMock()
+    mock_fn_7 = AsyncMock()
+    mock_fn_8 = AsyncMock()
+    callback_list = [
+        {
+            "function": mock_fn_1, "params": {"args": (1, 2), "kwargs": {"key": "value"}},
+            "inject_task_node": True,
+        },
+        {"function": mock_fn_2, "params": {"args": (1, 2), "kwargs": {"key": "value"}}},
+        {"function": mock_fn_3, "inject_task_node": True},
+        {"function": mock_fn_4},
+        {"function": mock_fn_5, "params": {"args": (1, 2), "kwargs": {"key": "value"}}, "inject_task_node": True},
+        {"function": mock_fn_6, "params": {"args": (1, 2), "kwargs": {"key": "value"}}},
+        {"function": mock_fn_7, "inject_task_node": True},
+        {"function": mock_fn_8},
+    ]
+    mock_callback.at_job_start = callback_list
+    mock_callback.at_job_end = []
+    setattr(mock_callback, "error_callback_type", None)
+    delattr(mock_callback, "error_callback_type")
+
+    # Assert
+    with pytest.raises(ValueError):
+        await commander._callback_handle(callback=mock_callback, which="error_callback_type")
+    
+    # Action
+    await commander._callback_handle(callback=mock_callback, which="at_job_end")
+    
+    # Assert
+    assert not mock_fn_1.called
+    assert not mock_fn_2.called
+    assert not mock_fn_3.called
+    assert not mock_fn_4.called
+    assert not mock_fn_5.called
+    assert not mock_fn_6.called
+    assert not mock_fn_7.called
+    assert not mock_fn_8.called
+
+    # Action
+    await commander._callback_handle(callback=mock_callback, which="at_job_start")
+
+    # Assert
+    mock_fn_1.assert_called_once_with(1, 2, key="value", task_node=mock_task_node)
+    mock_fn_2.assert_called_once_with(1, 2, key="value")
+    mock_fn_3.assert_called_once_with(task_node=mock_task_node)
+    mock_fn_4.assert_called_once_with()
+    mock_fn_5.assert_called_once_with(1, 2, key="value", task_node=mock_task_node)
+    mock_fn_6.assert_called_once_with(1, 2, key="value")
+    mock_fn_7.assert_called_once_with(task_node=mock_task_node)
+    mock_fn_8.assert_called_once_with()
+
+
+def test_call_handler_threadsafe(commander: CommanderAsync, job_add, handler_add):
+    # Setup
+    manipulate = [0]
+    job = job_add(manipulate)
+    handler = handler_add(manipulate)
+    
+    # Action
+    threading.Thread(target=commander.run, args=(job,)).start()
+    while not commander.running_status:
+        pass
+    commander.call_handler_threadsafe(handler)
+    while not commander.is_empty():
+        pass
+    commander.exit()
+
+    # Assert
     assert manipulate[0] == 2